Add Unicode arg parse for all utilites
