+++ src/blockchain_utilities/blockchain_blackball.cpp
@@ -167,6 +167,17 @@ static bool for_all_transactions(const std::string &filename, const std::functio
 
 int main(int argc, char *argv[])
 {
+#ifdef WIN32
+	std::vector<std::string> args;
+	std::vector<char*> argptrs;
+	command_line::set_console_utf8();
+	if(command_line::get_windows_args(args, argptrs))
+	{
+		argc = args.size();
+		argv = argptrs.data();
+	}
+#endif
+
 	TRY_ENTRY();
 
 	epee::string_tools::set_module_name_and_folder(argv[0]);
