+++ src/common/command_line.cpp
@@ -52,6 +52,11 @@
 #include <boost/algorithm/string/predicate.hpp>
 #include <unordered_set>
 
+#if defined(WIN32)
+#include <windows.h>
+#include <shellapi.h>
+#endif
+
 namespace command_line
 {
 namespace
@@ -89,6 +94,35 @@ bool is_no(const std::string &str)
 
 	return false;
 }
+
+#ifdef WIN32
+bool get_windows_args(std::vector<std::string>& args, std::vector<char*>& argptrs)
+{
+	int nArgs = 0;
+	LPWSTR* szArgs = CommandLineToArgvW(GetCommandLineW(), &nArgs);
+
+	if(szArgs == nullptr)
+		return false;
+
+	for(int i = 0; i < nArgs; i++)
+	{
+		int size_needed = WideCharToMultiByte(CP_UTF8, 0, szArgs[i], -1, NULL, 0, NULL, NULL);
+		args.emplace_back(size_needed, '\0');
+		std::string& str = args.back();
+		char* strptr = &str[0];
+		WideCharToMultiByte(CP_UTF8, 0, szArgs[i], -1, strptr, size_needed, NULL, NULL);
+		str.pop_back();
+		argptrs.emplace_back(strptr);
+	}
+
+	return true;
+}
+
+void set_console_utf8()
+{
+	SetConsoleCP(CP_UTF8);
+}
+#endif
 
 const arg_descriptor<bool> arg_help = {"help", "Produce help message"};
 const arg_descriptor<bool> arg_version = {"version", "Output version information"};
