+++ src/daemonizer/daemonizer.h
@@ -66,7 +66,7 @@ boost::filesystem::path get_relative_path_base(
    */
 template <typename T_executor>
 bool daemonize(
-	int argc, char const *argv[], T_executor &&executor // universal ref
+	int argc, char* argv[], T_executor &&executor // universal ref
 	,
 	boost::program_options::variables_map const &vm);
 }
