+++ src/daemonizer/windows_daemonizer.inl
@@ -67,7 +67,7 @@ const command_line::arg_descriptor<bool> arg_stop_service = {
 const command_line::arg_descriptor<bool> arg_is_service = {
 	"run-as-service", "Hidden -- true if running as windows service"};
 
-std::string get_argument_string(int argc, char const *argv[])
+std::string get_argument_string(int argc, char* argv[])
 {
 	std::string result = "";
 	for(int i = 1; i < argc; ++i)
@@ -129,7 +129,7 @@ inline boost::filesystem::path get_relative_path_base(
 
 template <typename T_executor>
 inline bool daemonize(
-	int argc, char const *argv[], T_executor &&executor // universal ref
+	int argc, char* argv[], T_executor &&executor // universal ref
 	,
 	boost::program_options::variables_map const &vm)
 {
